slack-bot-7d/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.3
pandas==1.4.1
pycoingecko==2.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
slack-sdk==3.15.2
tabulate==0.8.9
urllib3==1.26.9
